requirements.txt
boto3==1.34.1
botocore==1.34.1
numpy==1.26.2
pandas==2.1.3
plotly==5.18.0
psycopg2-binary==2.9.9
pyportfolioopt==1.5.5
python-dotenv==1.0.1
redis==5.0.1
regex==2023.10.3
requests==2.31.0
requests-file==1.5.1
s3transfer==0.9.0
scipy==1.11.4
seaborn==0.13.0
spacy==3.7.2
stocknews==0.9.11
streamlit==1.29.0
urllib3==2.0.7
yfinance==0.2.32
